+++ src/motor.h
@@ -24,14 +24,14 @@ class motor { //モーターチャンネルとピン設定
 };
 void motor::rotate()
 {
-  user real = {duty_min + abs(def + c1) * nf,
-              duty_min + abs(def + c2) * nf,
-              duty_min + abs(def + c3) * nf,
-              duty_min + abs(def + c4) * nf};
-  ledcWrite(ch1, real.x < def ? def : real.x);
-  ledcWrite(ch2, real.y < def ? def : real.y);
-  ledcWrite(ch3, real.z < def ? def : real.z);
-  ledcWrite(ch4, real.turn < def ? def : real.turn);
+  user real = {static_cast<double>(duty_min + abs(def + c1) * nf),
+              static_cast<double>(duty_min + abs(def + c2) * nf),
+              static_cast<double>(duty_min + abs(def + c3) * nf),
+              static_cast<double>(duty_min + abs(def + c4) * nf)};
+  ledcWrite(ch1, static_cast<int>(real.x) < def ? def : real.x);
+  ledcWrite(ch2, static_cast<int>(real.y) < def ? def : real.y);
+  ledcWrite(ch3, static_cast<int>(real.z) < def ? def : real.z);
+  ledcWrite(ch4, static_cast<int>(real.turn) < def ? def : real.turn);
 /*
   ledcWrite(ch1, duty_min + abs(def + c1) * nf);
   ledcWrite(ch2, duty_min + abs(def + c2) * nf);
